refactor: :recycle: move axios config in shared/http
