+++ src/api/register-auth.ts
@@ -1,4 +1,4 @@
-import { apiAuth } from './api-config';
+import { api } from '@/shared/http';
 
 // Работа с регистрацией и аутентификацией
 
@@ -18,17 +18,17 @@ type UserAuth = {
   roles: string[];
 };
 // регистрация пользователя
-export const registerUser = async (user: UserRegister): Promise<{ token: string }> => {
-  const response = await apiAuth.post<{ token: string }>('/auth/register', user);
-  return response.data;
+export const registerUser = async (user: UserRegister): Promise<number> => {
+  const response = await api.post('/auth/register', user);
+  return response.status;
 };
 
 // аутентификация пользователя
 export const authUser = async (user: UserLogin): Promise<{ token: string }> => {
-  const response = await apiAuth.post<{ token: string }>('/auth/authenticate', user);
+  const response = await api.post<{ token: string }>('/auth/authenticate', user);
   return response.data;
 };
 export const getCurrentUser = async (): Promise<UserAuth> => {
-  const response = await apiAuth.get<UserAuth>('/users/current');
+  const response = await api.get<UserAuth>('/users/current');
   return response.data;
 };
