+++ src/api/req-doc-attributes.ts
@@ -1,4 +1,4 @@
-import { api } from './api-config';
+import { api } from '@/shared/http';
 
 type Attribute = {
   id: number;
