+++ src/shared/http/index.ts
@@ -1,0 +1,1 @@
+export { api } from './axios-config';
